Move CircuitContext to runtime
