+++ circuit/src/commonMain/kotlin/com/slack/circuit/Presenter.kt
@@ -3,6 +3,7 @@
 package com.slack.circuit
 
 import androidx.compose.runtime.Composable
+import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Screen
 import kotlinx.coroutines.flow.MutableSharedFlow
