+++ circuit/src/commonMain/kotlin/com/slack/circuit/Ui.kt
@@ -4,6 +4,7 @@ package com.slack.circuit
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Screen
