+++ samples/counter/desktop/src/main/kotlin/com/slack/circuit/sample/counter/desktop/DesktopCounterCircuit.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.window.singleWindowApplication
 import com.slack.circuit.CircuitCompositionLocals
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitContent
-import com.slack.circuit.CircuitContext
+import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.Screen
 import com.slack.circuit.Ui
 import com.slack.circuit.sample.counter.CounterEvent
