+++ samples/counter/mosaic/src/main/kotlin/com/slack/circuit/sample/counter/mosaic/MosaicCounterCircuit.kt
@@ -17,7 +17,7 @@ import com.jakewharton.mosaic.runMosaic
 import com.slack.circuit.CircuitCompositionLocals
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitContent
-import com.slack.circuit.CircuitContext
+import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.Screen
 import com.slack.circuit.Ui
 import com.slack.circuit.sample.counter.CounterEvent
