+++ samples/interop/src/main/kotlin/com/slack/circuit/sample/interop/MainActivity.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.unit.dp
 import com.slack.circuit.CircuitCompositionLocals
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitContent
-import com.slack.circuit.CircuitContext
+import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.Navigator
 import com.slack.circuit.Presenter
 import com.slack.circuit.Ui
